@@ -6,11 +6,13 @@ import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
 import { HttpClientModule } from '@angular/common/http';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { TransactionId2StrPipe } from './total-transactions/custom.pipe';
 
 @NgModule({
-  declarations: [AppComponent],
+  declarations: [AppComponent,TransactionId2StrPipe],
   entryComponents: [],
-  imports: [BrowserModule, IonicModule.forRoot(), AppRoutingModule,    HttpClientModule,],
+  imports: [BrowserModule, IonicModule.forRoot(), AppRoutingModule,    HttpClientModule, BrowserAnimationsModule,],
   providers: [{ provide: RouteReuseStrategy, useClass: IonicRouteStrategy }],
   bootstrap: [AppComponent],
 })
