@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { DataService,Transactions } from '../services/data.service';
 
 @Component({
   selector: 'app-total-transactions',
@@ -6,10 +7,11 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./total-transactions.page.scss'],
 })
 export class TotalTransactionsPage implements OnInit {
-
-  constructor() { }
+  private transactions: Transactions[];
+  constructor(private data:DataService) { }
 
   ngOnInit() {
+    this.data.getTotalTransactions().subscribe(d => {this.transactions=d; console.warn('salam');});
   }
 
 }
